pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pynomina"
description = "Personal finance management utility"
keywords = ["finance", "money", "management","nicegui"]
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"},
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
]

dependencies = [
    # https://pypi.org/project/pylodstorage/
    "pylodstorage>=0.13.3",
    # https://pypi.org/project/networkx/
    "networkx>=3.3",
    # https://pypi.org/project/py-sidif/
    "py-sidif>=0.2.0",
    # NiceGUI widgets
    "ngwidgets>=0.19.2",
    # xsdata for XML data binding
    "xsdata>=24.9",
    # https://pypi.org/project/piecash/
    #"piecash>=1.2.1",
    #"SQLAlchemy==1.4.48",
    #"SQLAlchemy-Utils==0.37.2",
    # https://github.com/chardet/chardet
    "chardet>=5.2.0",
    # beancount
    # https://github.com/beancount/beancount
    # https://pypi.org/project/beancount/2.3.6/
    # was previous version
    "beancount>=3.0.0",
    # https://pypi.org/project/tabulate/
    "tabulate>=0.9.0",
    # beanquery
    # https://pypi.org/project/beanquery/
    "beanquery"

]

dynamic = ["version"]

[tool.hatch.version]
path = "nomina/__init__.py"


[project.urls]
Home = "https://github.com/WolfgangFahl/pynomina"
Documentation = "https://wiki.bitplan.com/index.php/pynomina"
Source = "https://github.com/WolfgangFahl/pynomina"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["nomina","nomina_examples"]

[tool.hatch.build.targets.wheel.sources]
"nomina" = "nomina"
"nomina_examples" = "nomina_examples"

[project.scripts]
nomina = "nomina.nomina_cmd:main"
money-zip = "nomina.money_zip:main"
npbeanquery = "nomina.beanquery_cmd:main"
